# src/api/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"AuthUrlResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"authUrl": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"authUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiError": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LinkLauncherBody": {
				"properties": {
					"launcherType": {
						"type": "string"
					},
					"accountName": {
						"type": "string"
					},
					"platformUserId": {
						"type": "string"
					}
				},
				"required": [
					"launcherType",
					"accountName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExchangeCodeBody": {
				"properties": {
					"code": {
						"type": "string"
					},
					"state": {
						"type": "string"
					}
				},
				"required": [
					"code",
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LocalConfigResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"config": {}
				},
				"required": [
					"success",
					"config"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LocalConfigBody": {
				"properties": {
					"installBasePath": {
						"type": "string"
					},
					"executablePattern": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Launcher of Launchers API",
		"version": "1.0.0",
		"description": "Launcher of launchers",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/launchers/linked": {
			"get": {
				"operationId": "GetLinkedAccounts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Devuelve las cuentas de launcher vinculadas del usuario",
				"tags": [
					"Launchers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/launchers/{launcherType}/auth-url": {
			"get": {
				"operationId": "GetAuthUrl",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthUrlResponse"
								}
							}
						}
					},
					"400": {
						"description": "Launcher no soportado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Devuelve la URL de autenticación OAuth2/OpenID para el launcher indicado.\r\nSteam usa OpenID 2.0; el resto usa OAuth2.",
				"tags": [
					"Launchers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "launcherType",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "returnUrl",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/launchers/{launcherType}/callback": {
			"get": {
				"operationId": "HandleCallback",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"400": {
						"description": "Parámetros inválidos",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Callback OAuth2 para launchers que redirigen al backend directamente.\r\nRuta pública — no requiere JWT.",
				"tags": [
					"Launchers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "launcherType",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "code",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/launchers/link": {
			"post": {
				"operationId": "LinkLauncher",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "Datos inválidos",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Vincula manualmente una cuenta de launcher (sin OAuth2)",
				"tags": [
					"Launchers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LinkLauncherBody"
							}
						}
					}
				}
			}
		},
		"/launchers/{accountId}": {
			"delete": {
				"operationId": "UnlinkLauncher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"404": {
						"description": "Launcher account not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Desvincula una cuenta de launcher",
				"tags": [
					"Launchers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "accountId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/launchers/{accountId}/sync": {
			"post": {
				"operationId": "SyncLauncher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"errors": {
											"type": "number",
											"format": "double"
										},
										"count": {
											"type": "number",
											"format": "double"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"errors",
										"count",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Linked launcher account not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Sincroniza un launcher específico del usuario.\r\nDevuelve el resumen de juegos agregados/actualizado.",
				"tags": [
					"Launchers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "accountId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/launchers/sync/all": {
			"post": {
				"operationId": "SyncAllLaunchers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"summary": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"summary",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Sincroniza todos los launchers vinculados del usuario.",
				"tags": [
					"Launchers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/launchers/scan/local": {
			"get": {
				"operationId": "ScanLocalInstallations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"summary": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"summary",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Realiza un escaneo local directo de instalaciones para todos los launchers que soportan detección de estado.\r\nNo requiere sincronización previa y no escribe a la base de datos.",
				"tags": [
					"Launchers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/launchers/{launcherType}/scan/local": {
			"get": {
				"operationId": "ScanLocalInstallationsByLauncher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"summary": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"summary",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Unable to scan",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Realiza un escaneo local directo solo para el tipo de launcher solicitado.",
				"tags": [
					"Launchers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "launcherType",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/launchers/{launcherType}/exchange": {
			"post": {
				"operationId": "ExchangeCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageResponse"
								}
							}
						}
					},
					"400": {
						"description": "code and state are required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Utilizado para launchers cuyo redirect_uri apunta a una URL externa (p.ej GOG).\r\nEl frontend extrae el `code` y `state` de esa URL de redirección y\r\nlos envía aquí para que el backend complete el intercambio OAuth2.",
				"tags": [
					"Launchers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "launcherType",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ExchangeCodeBody"
							}
						}
					}
				}
			}
		},
		"/launchers/{launcherType}/local-config": {
			"get": {
				"operationId": "GetLocalConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LocalConfigResponse"
								}
							}
						}
					}
				},
				"description": "GET /api/launchers/:launcherType/local-config\r\nDevuelve la configuración de instalación local almacenada para este launcher.",
				"tags": [
					"Launchers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "launcherType",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpsertLocalConfig",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LocalConfigResponse"
								}
							}
						}
					},
					"400": {
						"description": "At least one property is required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "PUT /api/launchers/:launcherType/local-config\r\nCrea o actualiza la configuración de instalación local para este launcher.\r\n\r\ninstallBasePath   — directorio raíz personalizado donde se instalan los juegos\r\nexecutablePattern — plantilla para la ruta del ejecutable; soporta\r\n                    placeholders {installPath} y {gameName},\r\n                    p.ej \"{installPath}\\\\bin\\\\{gameName}.exe\"",
				"tags": [
					"Launchers"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "launcherType",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LocalConfigBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}